refactor: move certificate generation to its controller
