+++ backend/src/Controllers/Project/ProjectDtos.cs
@@ -72,8 +72,3 @@ public class ProjectOperationSheetExport
     // personal_4
     public string Staff4 { get; set; } = string.Empty;
 }
-
-public class ProjectCertificateExport
-{
-
-}
